requirements.txt
yfinance
pandas
numpy
sqlalchemy
openpyxl
psycopg2-binary
python-dotenv
dotenv